Se borran carpetas que no se usan
