--- LicitAR.Web/Views/Account/Login.cshtml
@@ -1,73 +1,0 @@
-@page
-@{
-  Layout = "_BlankLayout";
-  ViewData["Title"] = "Login - LicitAR";
-}
-@model LicitAR.Web.Views.Account.Login
-@section PageStyles {
-<link rel="stylesheet" href="~/vendor/css/pages/page-auth.css">
-}
-
-@* ************** Content ************** *@
-
-<div class="container-xxl">
-  <div class="authentication-wrapper authentication-basic container-p-y">
-    <div class="authentication-inner">
-      <!-- Register -->
-      <div class="card px-sm-6 px-0">
-        <div class="card-body">
-          <!-- Logo - ->
-          <div class="app-brand justify-content-center">
-            <a href="/" class="app-brand-link gap-2">
-              <span class="app-brand-logo demo">@await Html.PartialAsync("../_Partials/_Macros")</span>
-              <span class="app-brand-text demo text-heading fw-bold">@TempData.Peek("appName")</span>
-            </a>
-          </div>
-          <!-- /Logo -->
-          <h4 class="mb-1">Welcome to @TempData.Peek("appName")! 👋</h4>
-          <p class="mb-6">Please sign-in to your account and start the adventure</p>
-
-          <form id="formAuthentication" class="mb-6" asp-action="Login" method="post">
-
-              
-            <div class="mb-6">
-              <label for="email" class="form-label">Email or Username</label>
-              <input type="text" class="form-control" id="email" name="email"required placeholder="Email" autofocus>
-            </div>
-            <div class="mb-6 form-password-toggle">
-              <label class="form-label" for="password">Password</label>
-              <div class="input-group input-group-merge">
-                <input type="password" id="password" class="form-control" name="password" placeholder="&#xb7;&#xb7;&#xb7;&#xb7;&#xb7;&#xb7;&#xb7;&#xb7;&#xb7;&#xb7;&#xb7;&#xb7;" aria-describedby="password" required/>
-                <span class="input-group-text cursor-pointer"><i class="bx bx-hide"></i></span>
-              </div>
-            </div>
-            <div class="mb-8">
-              <div class="d-flex justify-content-between mt-8">
-                <div class="form-check mb-0 ms-2">
-                  <input class="form-check-input" type="checkbox" id="remember-me">
-                  <label class="form-check-label" for="remember-me">
-                    Remember Me
-                  </label>
-                </div>
-                <a href="/Auth/ForgotPasswordBasic">
-                  <span>Forgot Password?</span>
-                </a>
-              </div>
-            </div>
-            <div class="mb-6">
-              <button class="btn btn-primary d-grid w-100" type="submit">Login</button>
-            </div>
-          </form>
-
-          <p class="text-center">
-            <span>New on our platform?</span>
-            <a href="/Auth/RegisterBasic">
-              <span>Create an account</span>
-            </a>
-          </p>
-        </div>
-      </div>
-      <!-- /Register -->
-    </div>
-  </div>
-</div>
