--- LicitAR.Web/Views/Account/Register.cshtml
@@ -1,14 +1,0 @@
-@page
-@model LicitAR.Web.Views.Account.Register
-@{
-    ViewData["Title"] = "Registro";
-}
-
-<h2>@ViewData["Title"]</h2>
-
- 
-<form asp-action="Register" method="post">
-    <input type="email" name="email" placeholder="Email" required />
-    <input type="password" name="password" placeholder="Password" required />
-    <button type="submit">Registrarse</button>
-</form>
