--- LicitAR.Web/Views/Account/View.cshtml
@@ -1,31 +1,0 @@
-﻿@model LicitAR.Core.Data.Models.Parametros.Localidad
-
-@{
-    ViewData["Title"] = "View";
-    Layout = "~/Views/Shared/_Layout.cshtml";
-}
-
-<h1>View</h1>
-
-<div>
-    <h4>Localidad</h4>
-    <hr />
-    <dl class="row">
-        <dt class = "col-sm-2">
-            @Html.DisplayNameFor(model => model.Descripcion)
-        </dt>
-        <dd class = "col-sm-10">
-            @Html.DisplayFor(model => model.Descripcion)
-        </dd>
-        <dt class = "col-sm-2">
-            @Html.DisplayNameFor(model => model.IdProvincia)
-        </dt>
-        <dd class = "col-sm-10">
-            @Html.DisplayFor(model => model.IdProvincia)
-        </dd>
-    </dl>
-</div>
-<div>
-    <a asp-action="Edit" asp-route-id="@Model?.IdLocalidad">Edit</a> |
-    <a asp-action="Index">Back to List</a>
-</div>
