+++ LicitAR.Web/Views/Shared/Sections/Navbar/_NavbarPartial.cshtml
@@ -68,7 +68,7 @@
           <div class="dropdown-divider my-1"></div>
         </li>
         <li>
-          <a class="dropdown-item" href="javascript:void(0);">
+          <a class="dropdown-item" asp-controller="Usuario" asp-action="MisDatos">
             <i class="bx bx-user bx-md me-2"></i>
             <span>Mi perfil</span>
           </a>
