+++ LicitAR.Web/Views/Usuario/Login.cshtml
@@ -1,9 +1,10 @@
-﻿  
+﻿@using LicitAR.Web.Views.Usuario
+  
 @{
   Layout = "_BlankLayout";
   ViewData["Title"] = "Login - LicitAR";
 }
-@model LicitAR.Web.Views.Account.Login
+@model Login
 @section PageStyles {
     <link rel="stylesheet" href="~/vendor/css/pages/page-auth.css">
 }
