+++ LicitAR.Web/Views/Usuario/MisDatos.cshtml
@@ -1,0 +1,79 @@
+﻿@model LicitAR.Web.Business.Identidad.Usuario.RegistroModel
+
+@{
+    ViewData["Title"] = "Mis Datos";
+    Layout = "~/Views/Shared/_Layout.cshtml";
+}
+
+<h1>View</h1>
+
+<div>
+    <h4>Localidad</h4>
+    <hr />
+    @using (Html.BeginForm("Register", "Usuario", FormMethod.Post, new { @class = "needs-validation", novalidate = "novalidate" }))
+    {
+        @Html.AntiForgeryToken()
+
+        <div class="row">
+            .
+            <div class="col-md-12 mb-3">
+                @Html.LabelFor(m => m.Cuit, new { @class = "form-label" })
+                @Html.TextBoxFor(m => m.Cuit, new { @class = "form-control", maxlength = 11 })
+                @Html.ValidationMessageFor(m => m.Cuit, "", new { @class = "text-danger" })
+            </div>
+            <div class="col-md-12 mb-3">
+                @Html.LabelFor(m => m.Nombre, new { @class = "form-label" })
+                @Html.TextBoxFor(m => m.Nombre, new { @class = "form-control", maxlength = 30, required = "required" })
+                @Html.ValidationMessageFor(m => m.Nombre, "", new { @class = "text-danger" })
+            </div>
+
+            <div class="col-md-12 mb-3">
+                @Html.LabelFor(m => m.Apellido, new { @class = "form-label" })
+                @Html.TextBoxFor(m => m.Apellido, new { @class = "form-control", maxlength = 30, required = "required" })
+                @Html.ValidationMessageFor(m => m.Apellido, "", new { @class = "text-danger" })
+            </div>
+
+            <div class="col-md-12 mb-3">
+                @Html.LabelFor(m => m.FechaNacimiento, new { @class = "form-label" })
+                @Html.TextBoxFor(m => m.FechaNacimiento, "{0:yyyy-MM-dd}", new { @class = "form-control", type = "date", required = "required" })
+                @Html.ValidationMessageFor(m => m.FechaNacimiento, "", new { @class = "text-danger" })
+            </div>
+
+        </div>
+
+
+        <h5 class="py-3 mt-4"><i class="fas fa-user-plus  "></i> Datos de Usuario</h5>
+
+        <div class="row">
+
+            <div class="col-md-12 mb-3">
+                @Html.LabelFor(m => m.Email, new { @class = "form-label" })
+                @Html.TextBoxFor(m => m.Email, new { @class = "form-control", maxlength = 50, required = "required" })
+                @Html.ValidationMessageFor(m => m.Email, "", new { @class = "text-danger" })
+            </div>
+
+            <div class="col-md-12 mb-3">
+                @Html.LabelFor(m => m.Password, new { @class = "form-label" })
+                @Html.PasswordFor(m => m.Password, new { @class = "form-control", maxlength = 50, required = "required" })
+                @Html.ValidationMessageFor(m => m.Password, "", new { @class = "text-danger" })
+            </div>
+
+            <div class="col-md-12 mb-3">
+                @Html.LabelFor(m => m.ConfirmaPassword, new { @class = "form-label" })
+                @Html.PasswordFor(m => m.ConfirmaPassword, new { @class = "form-control", maxlength = 50, required = "required" })
+                @Html.ValidationMessageFor(m => m.ConfirmaPassword, "", new { @class = "text-danger" })
+            </div>
+
+        </div>
+
+        <div class="text-end">
+            <button type="submit" id="btnSubmit" class="btn btn-primary">Registrar</button>
+        </div>
+    }
+</div>
+
+</div>
+<div>
+    <a asp-action="Edit" asp-route-id="@Model?.IdUsuario">Edit</a> |
+    <a asp-action="Index">Back to List</a>
+</div>
